Telehealth and Virtual Hospitals lab
this lab focuses on the use of the ESP8266 microcontroller NodeMCU to host a web page to serve an interactive page allowing us to control LEDs over Wi-Fi. Example of this application can be seen in the remote control of hospital room lights by care givers; this example presents a button that sends a post request to the LED causing it to come on and off.